+++ CHANGELOG.md
@@ -12,12 +12,15 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 - Added schedule (cron string) to orbitjob CRD
 - Added monitoring to OrbitJob so pytest can evaluate the jobs execution
 - Added PyTest for lake admin, creator and user
+- Added 'one-click' support for trial usage of AWS Workbench
+
 ### **Changed**
 
 - Moved away from Job index to Status object of K8s job to report the current status
 - Updated SDK to use OrbitJobs instead of Jobs and CronJobs
 - Changed Lake Creator notebook
 - Added sleep to Orbit SDK based OrbitJob CR job status check
+- Fixed the expiring kubeconfig issue for long living k8s jobs
 
 ### **Removed**