+++ CHANGELOG.md
@@ -20,6 +20,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 - Updated SDK to use OrbitJobs instead of Jobs and CronJobs
 - Changed Lake Creator notebook
 - Added sleep to Orbit SDK based OrbitJob CR job status check
+- Fixed the expiring kubeconfig issue for long living k8s jobs
 
 ### **Removed**