+++ CHANGELOG.md
@@ -54,6 +54,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 - FIX: regression tests for DataBrew needed parameterized names
 - FIX: fix podsettings filter when copying into user namespace
 - FIX: imagereplication-pod-webhook patching when no changes occur
+- FIX: Removed the istio-system PDB before destroying istio-system resources, causing dead-lock
 
 ### **Removed**